// src/schemas/generated/types/designStyle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HomeDecorAI API specification
 * OpenAPI spec version: 0.1.0
 */

export type DesignStyle = (typeof DesignStyle)[keyof typeof DesignStyle];

export const DesignStyle = {
  modern: "modern",
  minimalist: "minimalist",
  scandinavian: "scandinavian",
  industrial: "industrial",
  bohemian: "bohemian",
  contemporary: "contemporary",
  midCentury: "midCentury",
  coastal: "coastal",
  farmhouse: "farmhouse",
  japandi: "japandi",
  artDeco: "artDeco",
  traditional: "traditional",
  tropical: "tropical",
  rustic: "rustic",
  luxury: "luxury",
  cozy: "cozy",
  christmas: "christmas",
  airbnb: "airbnb",
} as const;